+++ pycaption/scc/__init__.py
@@ -1,3 +1,5 @@
+"""SCC (Scenarist Closed Captions) reader and writer for CEA-608 caption data."""
+
 from .reader import SCCReader
 from .writer import SCC_TOKENS_PER_CAPTION_MAX, SCCWriter